appraisals support & rails 4.0 support
