+++ Appraisals
@@ -2,6 +2,14 @@ appraise "rails-4" do
   gem "rails", "4.0.0"
 end
 
+appraise "rails-4.0.3" do
+  gem "rails", "4.0.3"
+end
+
+appraise "rails-4.1" do
+  gem "rails", "4.1.0"
+end
+
 appraise "rails-4.2" do
   gem "rails", "4.2.0"
 end
