+++ gemfiles/rails_4.0.3.gemfile
@@ -1,0 +1,20 @@
+# This file was generated by Appraisal
+
+source "https://rubygems.org"
+
+gem "certified"
+gem "debugger", :platform => :ruby_19
+gem "pry"
+gem "pry-remote"
+gem "pry-nav"
+gem "sqlite3"
+gem "appraisal"
+gem "rails", "4.0.3"
+
+group :test do
+  gem "rspec"
+  gem "vcr"
+  gem "webmock", "1.16.0", :require => false
+end
+
+gemspec :path => "../"
